Got rid of some compilation warnings in gui

+++ HOPSAN++/HopsanGUI/UndoStack.cpp
@@ -402,7 +402,7 @@ void UndoStack::undoOneStep()
         }
     }
 
-    for(size_t i=0; i<addedWidgetList.size(); ++i)
+    for(int i=0; i<addedWidgetList.size(); ++i)
     {
         mpParentContainerObject->mWidgetMap.find(addedWidgetList.at(i)).value()->deleteMe(NOUNDO);
     }
